Initial commit of a semi working twikrad.
twikrad is a hacked up port of wikrad, the command line
socialtext editor. twikrad does the same thing for
tiddlyweb.

It is not yet mature, but seems to do the basics.

This is a hacked up port because rather than trying to find
some common ground and extract some abstractions, which proved
too hard, this just copies all the code into a new package
namespace and hacks and slashes to get things working.

Socialtext::Resting, Socialtext::Resting::Utils, etc have
all been worked into TiddlyWeb:: in lib in this repo. This
will be tidied up with time.